refactor: upload.js 코드 리펙토링
